discovered hibernate problems solved

+++ src/main/java/com/wildiskin/cinema/models/Director.java
@@ -5,6 +5,7 @@ import jakarta.persistence.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.UUID;
 
 @Entity
 @Table(name="director")
@@ -44,7 +45,7 @@ public class Director {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
